cleaned up code. created baseline component for MVP UI. changed searchForDomain function to work with getIpData better.
cleaned up the code to be more logical and arguably to have better readability. current state of the application is in a play testing state. refactored helper functions to use async and await instead of callbacks and changed some variable names to create better readability. constructed baseline component for the data view to render MVP UI.